+++ src/components/GroupedBarChart/tests/GroupedBarChart.test.tsx
@@ -32,21 +32,9 @@ describe('GroupedBarChart />', () => {
     expect(groupedBarChart).toContainReactComponent(Legend);
   });
 
-  it('renders a <Chart /> if container dimensions have been measured', () => {
-    jest.useFakeTimers();
+  it('renders a <Chart />', () => {
     const groupedBarChart = mount(<GroupedBarChart {...mockProps} />);
-
-    groupedBarChart.act(() => {
-      jest.runAllTimers();
-    });
 
     expect(groupedBarChart).toContainReactComponent(Chart);
-    jest.useRealTimers();
-  });
-
-  it('does not render a <Chart /> if container dimensions have not been measured', () => {
-    const groupedBarChart = mount(<GroupedBarChart {...mockProps} />);
-
-    expect(groupedBarChart).not.toContainReactComponent(Chart);
   });
 });