+++ src/components/LineChart/LineChart.tsx
@@ -1,4 +1,4 @@
-import React, {useEffect, useRef, useState} from 'react';
+import React, {useLayoutEffect, useRef, useState} from 'react';
 import {useDebouncedCallback} from 'use-debounce';
 
 import {Chart} from './Chart';
@@ -29,9 +29,9 @@ export function LineChart({
     }
   }, 100);
 
-  useEffect(() => {
+  useLayoutEffect(() => {
     if (containerRef.current != null) {
-      updateDimensions();
+      setChartDimensions(containerRef.current.getBoundingClientRect());
     }
 
     window.addEventListener('resize', updateDimensions);