+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `<LineChart />` prop improvements to rename `x` to `label` and `y` to `rawValue`
 - `<LineChart />` displays label value on tooltip rather than Series name
 - `<LineChart />` returns null when empty data is passed in
+- Reduces time before chart components are rendered
 
 ## [0.0.15] - 2020-09-10